I'm still trying to make it work, in my new folder. I tried 'npm cache clean --force', then 'npm uninstall -g create-react-app', then 'npm install -g create-react-app'. That seemed to do something. Now it says 'Something is already running on port 3000'. I tried to see if there's something I can change in the code to change the port manually, but I'm not sure.
So I'm just responding to 'Would you like to run the app on another port instead?' with Y.  I think it changed to 3001. React app seems to be running! It's not letting me commit, though. It keeps loading. What if I... npm stop? ctrl+C? That didn't help!!! Ohhh, there are no changes to commit.

So, the first thing I need is img_data.txt, right? And read_data.py? And then create_table.py. This is where I might need to make some changes? Since it was being weird last time?

So... like what is even going on here? As far as I understand, I just want to make a database and put it in my bin. Right? But it looks like what it's doing is... using a host? And a port? And a database called images? But then I'm creating a database called img_table. Kind of confusing.

Hey, I committed changes, but it's still loading!!! Wat is this??? What if I force it to commit anyway? I hope I don't lose everything! I hope I didn't mess it up my forcing it to clean the cache... oh well, at least I have everything written down, right?