Oh, maybe I should get my MariaDB stuff out of the way... But how do I change the default port??? omg!!! I kind of hate to uninstall it totally because MySQL is so uncooperative!!! Well, let me try. The good thing about MariaDB is it was easy to install.
OK, it looks like I have a database in MySQL already, but how about I try dynamically creating a database here?

I tried navigating to http://127.0.0.1:5000/resetdb, and it was all like 'Not Found
The requested URL was not found on the server. If you entered the URL manually please check your spelling and try again.'! I've certainly seen that error before!