I tried doing the routes in server.py? But now I want to first of all plug in, like, a reset db button or something, connected to one of the routes, to test if they're working. We must have done that in an example!!!
I pasted a button in from an example, but now 'Cannot read properties of undefined (reading 'bind')'. So where's bind from again??? No, my imports are fine. Oh, my resetdb function in index.js is commented out!!! So I want to connect that to the one in server.py... How to connect to a function in different file???

I wrote the import, like, how I want to do it!!!