Maybe I'm looking for a fetch request. Or, I guess GET is the http request I'm looking for.
Oh, here's a person using Axios to GET something. So... that was like that onLoginChange thing, which I never got!!! Like, what was that??? I pasted it back in, whatever it was.

OK, I tried replacing onLoginChange with this resetdb function that uses axios.get, and renaming it getnext.

So, it's trying to get something out of the Python? Can't I get it to return one of its random datas, at least?

Now, it seems like it should be alerting with random decades (although it's not)

Oh wait, one last thing, there are two getnext's now, I should probably um, change that... I frankly don't get the whole config thing. Why is it using that, instead of just going axios.get? Well, I'm deleting it for now.

I'm still getting nexted, but now I'm getting 'AxiosError: Network Error's!