FIXED ALERT (correct mysql login credentials)
next, i'm going to try taking this get request and making it so it puts things in my state. and i got that state stuff basically working before, right? shouldn't be too hard?

so, what i ran into last, is i couldn't get the images to work. hopefully redoing it with mysql helps.