CREATETABLE/NEXTPHOTO (UNTESTED)
How about I separate createtable and nextphoto? Does that work? Logging in twice? Do I even have a good createtable going? Because... well, I do appear to have stuff in my table now, that's good.

But... see, the thing is, in the example, did you have it recreating the database every time? That doesn't seem like it would be efficient. But I guess it works, technically. But look, if I have a table with stuff, can't I just use that?

I predicted this would do that weird CORS error again, but now that I've reshuffled my functions in server.py,  it's doing an Axios 404 on createdb, and nothing at all on nextphoto. Oh wait, I think I put in the correct variable names and now it's giving me a 500 instead of a 404.