Today I'm trying to just paste sections of the code into Youtube and see what I find. First one I found was 'API Tutorial - Fetch Data with Axios', where I saw they got variables out of axios, which seems like it would be useful. On that video they were coding in App.js, with like, a function App() Instead of a class Main. I don't really get the difference??? Why are we using class Main??? So I didn't copy down all the formatting, but I at least tried getting a variable out of Axios. So I tried that with {success}, but it's probably still not working...
Well, hypothetically this could be useful in the future! I'm not quite doing it right. Still, shouldn't I be able to figure this out using just the alert? I will commit this and then go on to a few other video tutorials, which were maybe related to this one, or maybe I Youtubed the other snippet of Axios.