MESS OF FILENAME
OK, so... problem 1? Since maybe if I set up the event handler first then it will help with nextphoto? Uh, where was I with this? It's just... not working, right?

Now, here is the point where I do think it might be helpful to narrow down the variables.

I got the Python to the point where it's not showing any errors, but the next photo button still isn't working, and reset db gives error 500. Again? Did I ever get rid of that in the first place?

Anyway, this commit is a mess but let me just put it down and then maybe clear my head a little bit.