BOTCHED DEFLASKIFICATION, FROZEN OS
I'm just going to comment out the parts with app stuff, so I can really work on this python file.

Trying to disentangle this thing is challenging... now it says '  File "<frozen os>", line 679, in __getitem__
KeyError: 'DATAHOST''. So... is it using the flask app system to connect to the database? Well, OK, seems like commenting out all the app stuff might not be a viable plan...

What??? I think I uncommented all the stuff I had commented, and it's STILL giving me that error! So... is it because of the stuff I commented in check()?

The only changes left are, like, changes with the returns! Is that giving me a weird error?