STARTING TO JSONIFY
But wait, I might have to try jsonifying first, because probably the reason there's a server error on check , is because I'm not passing it what it wants! Sooo... what shows up on the screen when I return jsonify for nextphoto?

There!!! It printed something that looks like a JSON!!! Now... the part that seems strange, is like, where does the jsonify go after that? Don't you want to add it to index.js? Do I need to go pick it up somewhere? I guess I'd better watch the next part of the tutorial!!! I will save this now, because technically progress!