MYSTERIOUS LAG, CHECK ERROR REMAINS
Well, OK, I guess I'd better just try running this js! It seems to be basically working, but it's really laggy!!! Or, rather, it seems like I need to wait a couple seconds, after clicking it once, before I can click the Next Photo button again and it works. But that's not a reliable behavior, sometimes it randomly freezes for longer. And there are times when, like, after I've left the interface sitting idle for a while, when I go back in and click Next Photo, it doesn't respond at first, and/or then it spazzes out and when it updates, the whole interface goes lower on the screen for a moment.

It's also still not displaying the photo, and I still need to implement the actual program checking functionality. Although, since the jsonify seems to be working, I wonder if I would have more success with the check route now... let me check in my python... nope, that's still internal server error, so I'll have to figure that out.

Anyway, is it laggy because of all the data? I hope not! The fun apps are the ones with all the data!!! So, hopefully there is a reason for the lagginess that is easier to fix... I am just going to ignore it for now...