STRIP STRINGS IMAGES IN FLASK
omg!!!! I stripped decade and there's still a /n at the end of it!!! Why is it doing that??? Ohhh, I guess because it's reading from a text file.

This person is using rstrip, which strips on the right, which I guess makes sense. rstrip didn't work, but for some reason strip worked fine.

Now what about images in flask??? Oh, what a bunch of stuff. I'll look at it tomorrow!!! At least I fixed one of my problems!!!