MISSING PARENT ELEMENT
OK, I haven't looked at this in a while!!! Last time, download_file was a problem... I'm just commenting it out for now, I guess? Let me see if I get any more errors. And then I still need to get it so it's actually displaying the photo.

OK, it looks like it uploaded, here's where I get to my problem displaying images in React. I tried to go, like, '<img src = {address(filename)}...>' in my index.js, but that didn't work!!! Hopefully render_templates or whatever works better. The thing is, this example doesn't really show how to dusplay the photo. And I'm also confused about how all the stuff gets from newapp.py to imdex.html!!!

The thing is, like, I'm glad it's in the upload folder now, but how to you get THAT SPECIFIC PHOTO, corresponding to whatever photo is currently selected? As in, like, address(filename)?

Can I do it, through... def upload_file returns render_template('index.html'), right? Actually, that looks like it will be useful. Was that even in the other examples I looked at???

OK, so if I just use that, with what was in my index.js, then it displays the correct photo, on a different page... how do I get it to display the photo on the same page as everything else??? eg integrate with my React???

Well, it's trying to display an image. On its own page. I have it also doing that in index right now.

So, what about 'render template react'??? Is that even a thing? I guess I just need to look up how to display images, like, on the go? What's the word? Dynamically? I mean... was the little address part messing it up? That wasn't it, was it?

Oh, here's someone creating a path to the image! That sounds promising!

I just tried something like that, and tried to put in an f string, and now it's complaining about 'JSX expressions must have one parent element.'