IMAGE HOSTING?
Displaying/storing images in React...

1. Use the import keyword
I don't think that's going to work for all the images!!!

However, I'm going to try it with null.jpg... Yeah, it's working! There's also a require keyword.

Can I just put my images in the public folder?

See, I tried that, but it's just not working!!! I'm going to put them back for now, so there's not a bunch of changes...

Uh... I guess one option is just to host images externally and then put in the urls... I might do that, if I can't find anything else.

So, are there better alternatives to image hosting sites???